requirements.txt
Django~=5.1.1

openai~=1.50.2

gradio_client~=1.3.0
transformers>=4.39.0,<4.41.0
sentencepiece
descript-audio-codec
fastapi~=0.111.0
uvicorn~=0.29.0
pydantic-settings~=2.2.1
huggingface-hub==0.23.1
openai==1.50.2

djangorestframework~=3.15.2
requests~=2.32.3
